Got rid of a warning.

+++ Lec/Lec/ViewModels/LECHomeViewModel.m
@@ -27,7 +27,7 @@
 
 -(void)updateCourseIndexes{
     
-    int indexCount = self.tableData.count-1;
+    long indexCount = self.tableData.count-1;
     for(int i = 0; i < self.tableData.count; i++) {
         LECCourseCellViewModel *courseCell = [self.tableData objectAtIndex:indexCount-i];
         Course *course = courseCell.course;
